arm-riscos-gnueabi: set up GOT if there is one
Non-PIC (static) binaries may still link against PIC library code
and so require the GOT to be discoverable. Now that it is expected
to be found via the runtime workspace entry at &8038, we must
ensure that static binaries converted to AIF using elf2aif behave
correctly.

Unfortunately, this scheme will fallover if ever the OS makes the
text segment read-only (as &8038 is in the middle of the AIF
header, so cannot have GOTT_BASE baked in at conversion time).