arm-riscos-gnueabi: dump abort block and disassembly
If we're generating a traceback as the result of an environment
handler being invoked, there may well be a stored abort block
to inspect. Dump the registers it contains and disassemble the
code at the point of explosion, as for non-EABI builds.

Note, however, that at present, ARMEABISupport gobbles data aborts
so they end up appearing via the Error handler and thus no abort
block is available. This does mean that invalid data accesses no
longer produce useful stack traces.