arm-riscos-gnueabi: rework GOT base injection
The previous approach, as noted, would fail if the OS marked the
ROdata section readonly and also relied on collusion between
elf2aif and UnixLib. Instead, take a different approach which
removes the collusion and which will also work if the ROdata
section is not writeable: append relocation code to the end of
the image and ensure it is run before the zero-initialisation
code runs. By design (and the AIF specification), the ROdata
section _must_ be writeable when the relocation code runs, so
we can guarantee that it is safe to write things from it.

One wrinkle is that DDT makes use of the same location as the
optimised "well-known" GOTT_BASE location (&8038). There is an
unsatisfactory workaround for this scenario.

Note also that we are exceptionally tight on space for the
entrycode (in fact, if there are fewer than 3 program segments
in the ELF image, there are only 8 bytes remaining after the AIF
header, which is barely sufficient for anything). Fortunately,
we know this toolchain will emit at least 3 program headers so
this scheme is safe enough. If we ever need more space for the
entrycode, the obvious solution is to have the various
implementations of _start (found in the crt/*crt*.s files in
the UnixLib sources) begin with sufficient NOPs for elf2aif
to inject the entrycode into.