arm-riscos-gnueabi: make DDT hack less invasive
Lose 2 instructions (or 25%) from the entrycode by injecting
PC-relative instructions to compute the address of the GOT pointer
and store it into the "well known" GOTT_BASE at &8038.

This does make the relocation code somewhat larger, but there is
no space limitation there, so it's less likely to break. (The
requirement for at least 3 ELF program headers still remains as,
even with this reduction in entrycode size, it's still 16 bytes
larger than the space that would be available with 2 program
headers).