m68k-unknown-amigaos: pin a known-working binutils
It turns out that the Makefile would just fetch whatever was at
the top of the SVN branch, which is unfortunate. It also turns
out that nothing happens if there's already a stamp file
indicating that the sources have been obtained. Pin the revision
last fetched by the CI worker (which dates from 2017, so is
hardly modern) as that is known to work (and our patches apply
to it).