m68k-unknown-amigaos: disable fatal warnings in GCC14
The sources being built here are ancient and give GCC14 a large
amount of (reasonable) indigestion. It is intractable to patch
these, so simply stop the fatal ones being fatal. Trying to build
GCC 3.4.6 (or binutils 2.14) from source with any modern compiler
is increasingly unlikely to work without copious warnings, at
minimum.

Thus, this entire toolchain is in need of updating to something
more modern (if, indeed, such a thing exists).