Tools: add tooling to ingest upstream sources
This requires that pristine-tar 1.50 or later is installed, but
we carry our own copy of the top level script for this which has
local changes to minimise the delta for POSIX tarballs with pax
extensions (as produced for GCC releases, for example).

It also means that all source archives must now be tar format,
so upstream artefacts released as zip or lha files must be
manually repacked before import.